Incl effective date in ch_info sort when dropping dupes
The `kind` parameter of `sort_values` defaults to "quicksort", which is not stable.
We could specify a stable sort since the data is loaded in date order, but adding the "Effective Date:" column to the sort makes the sorting explicit, and pandas uses a stable sort when sorting by more than one column.